Update Elastic Search Indexer.

+++ CloudFabric.Libraries.Search.Indexer.ES/ElasticSearchIndexer.cs
@@ -46,31 +46,41 @@ namespace CloudFabric.Libraries.Search.Indexer.ES
                     string newIndexVersionSuffix = DateTime.Now.ToString("yyyyMMddHHmmss");
 
                     newIndexName = indexName + "-" + newIndexVersionSuffix;
+
+                    var response = await _client.IndexExistsAsync(
+                        new IndexExistsRequest(newIndexName)
+                    );
+                    if (response.Exists)
+                    {
+                        await _client.DeleteIndexAsync(
+                            new DeleteIndexRequest(newIndexName)
+                        );
+                        Console.WriteLine($"Deleted index " + newIndexName);
+                    }
+
+                    var descriptor = new CreateIndexDescriptor(newIndexName)
+                        .Settings(s => s
+                            .Analysis(analysis => analysis
+                            .Analyzers(analyzers => analyzers
+                                .Custom("folding-analyzer", c => c
+                                    .Tokenizer("standard")
+                                    .Filters("lowercase", "asciifolding")
+                                )
+                            )
+                        ));
+
+                    await _client.CreateIndexAsync(descriptor);
+                    Console.WriteLine($"Created index " + newIndexName);
                 }
 
-                await _client.DeleteIndexAsync(
-                    new DeleteIndexRequest(newIndexName)
-                );
+                var properties = GetPropertiesDescriptors<T>();
 
-                Console.WriteLine($"Deleted index " + newIndexName);
-
-                var descriptor = new CreateIndexDescriptor(newIndexName)
-                    .Settings(s => s
-                        .Analysis(analysis => analysis
-                        .Analyzers(analyzers => analyzers
-                            .Custom("folding-analyzer", c => c
-                                .Tokenizer("standard")
-                                .Filters("lowercase", "asciifolding")
-                            )
-                        )
-                    ))
-                    .Map(d => d
-                        .Properties<T>(p => GetPropertiesDescriptors(p))
-                    );
-
-                await _client.CreateIndexAsync(descriptor);
-
-                Console.WriteLine($"Created index " + newIndexName);
+                var putMappingRequest = new PutMappingRequest(newIndexName)
+                {
+                    Properties = properties
+                };
+                await _client.MapAsync(putMappingRequest);
+                Console.WriteLine($"Updated mapping for index " + newIndexName);
             }
             catch (Exception ex)
             {
@@ -80,8 +90,10 @@ namespace CloudFabric.Libraries.Search.Indexer.ES
             return true;
         }
 
-        private IPromise<IProperties> GetPropertiesDescriptors<T>(PropertiesDescriptor<T> properties) where T : class
+        private IProperties GetPropertiesDescriptors<T>() where T : class
         {
+            var properties = new PropertiesDescriptor<T>();
+
             PropertyInfo[] props = typeof(T).GetProperties();
             foreach (PropertyInfo prop in props)
             {
@@ -156,7 +168,7 @@ namespace CloudFabric.Libraries.Search.Indexer.ES
                 }
             }
 
-            return properties;
+            return ((IPromise<IProperties>)properties).Value;
         }
     }
 }
